Refactor isomorphic tests in core from mocha/enzyme to vitest/react-dom
