+++ packages/core/package.json
@@ -42,7 +42,7 @@
         "lint-fix": "es-lint --fix && sass-lint --fix",
         "prepack": "run-s compile dist",
         "test": "run-s test:iso test:vitest:run",
-        "test:iso": "mocha test/isotest.mjs",
+        "test:iso": "vitest run --config vitest.config.isotest.mts",
         "test:vitest": "vitest",
         "test:vitest:run": "vitest run",
         "verify": "npm-run-all compile -p dist test lint"
@@ -82,7 +82,6 @@
         "@vitejs/plugin-react": "catalog:",
         "enzyme": "catalog:",
         "jsdom": "catalog:",
-        "mocha": "^10.2.0",
         "npm-run-all": "^4.1.5",
         "react": "^18.3.1",
         "react-dom": "^18.3.1",
