+++ packages/test-commons/src/index.ts
@@ -15,5 +15,6 @@
 
 export * from "./datetimeUtils.js";
 export * from "./generateIsomorphicTests.js";
+export * from "./generateIsomorphicTestsVitest.js";
 export * from "./testErrorBoundary.js";
 export * from "./utils.js";
